Rename jobScraper.ts to hnhiringScraper.ts
